[Ordonnancement][POM] Regroupement des fichiers et mise en place du pom.xml
