+++ src/main/java/basNiveau/GestionnaireClientBas.java
@@ -1,4 +1,12 @@
-package basNiveau;
+/**
+ * Projet Programmation 2 - Licence professionnel SIL en alternance
+ * 14 décembre 2015 / 15 janvier 2016
+ *  
+ * @author Cadorel Maël - Blin Marina
+ * @version 1.0.0
+ */
+
+package main.java.basNiveau;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -7,8 +15,9 @@ import java.io.PrintWriter;
 import java.net.Socket;
 
 /**
- * 
- *
+ * Thread pour représenter une connexion d'un client
+ * au serveur echo. Il implémente la tâche de renvoyer
+ * au client ses messages.
  */
 public class GestionnaireClientBas implements Runnable
 {
@@ -18,13 +27,22 @@ public class GestionnaireClientBas implements Runnable
 	GestionnaireClientBas(Socket client, ServeurBas server)
 	{
 		this.client = client;
+		
+		// utile pour comptabiliser les déconnexions
 		this.server = server;
 	}
 
+	/**
+	 * Définit la tâche assignée à ce thread.
+	 * C'est le point d'entrée du thread, la 
+	 * méthode appelée lorsqu'on lance 
+	 * GestionnaireClientBas.start().
+	 */
 	public void run()
 	{
 		try
 		{
+			// chaque gestionnaire de client doit avoir son propre writer
 			PrintWriter writer = new PrintWriter(client.getOutputStream(), true);
 		
 			BufferedReader reader = new BufferedReader(new InputStreamReader(client.getInputStream()));
@@ -35,15 +53,19 @@ public class GestionnaireClientBas implements Runnable
 			{
 				writer.println("[Vous] : ");
 				String line = reader.readLine();
+				
+				// 'logging' basique sur le serveur
 				System.out.println("[Mot/Phrase à retourner] : " + line);
+				
 				if (line.trim().equals("\\quit"))
 				{
 					writer.println("A bientôt !");
 					break;
 				}
+				
+				// fonctionnalité d'echo
 				writer.println("[echo] : " + line);
 			}
-
 		}
 		catch (IOException e)
 		{
@@ -53,6 +75,7 @@ public class GestionnaireClientBas implements Runnable
 		{
 			try
 			{
+				// informe le serveur qu'une 'place' s'est libérée
 				server.removeClient();
 				client.close();
 			}
