+++ src/main/java/hautNiveau/GestionnaireClientHaut.java
@@ -1,4 +1,12 @@
-package hautNiveau;
+/**
+ * Projet Programmation 2 - Licence professionnel SIL en alternance
+ * 14 décembre 2015 / 15 janvier 2016
+ *  
+ * @author Cadorel Maël - Blin Marina
+ * @version 1.0.0
+ */
+
+package main.java.hautNiveau;
 
 import java.io.BufferedReader;
 import java.io.IOException;
