+++ src/main/java/hautNiveau/ServeurHaut.java
@@ -1,4 +1,12 @@
-package hautNiveau;
+/**
+ * Projet Programmation 2 - Licence professionnel SIL en alternance
+ * 14 décembre 2015 / 15 janvier 2016
+ *  
+ * @author Cadorel Maël - Blin Marina
+ * @version 1.0.0
+ */
+
+package main.java.hautNiveau;
 
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -8,16 +16,30 @@ import java.util.Properties;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+
+/**
+ * Implémentation haut niveau du serveur TCP echo.
+ * Renvoie les messages reçus via une connexion ServerSocket.
+ * Les threads sont gérés par un ExecutorService.
+ * 
+ */
 public class ServeurHaut
 {
 	Properties prop = new Properties();
 	private ServerSocket connexion;
 
+	/**
+	 * Constructeur. Il se charge de lire
+	 * le fichier de configuration et d'initialiser
+	 * le socket serveur. 
+	 */
 	public ServeurHaut()
 	{
 		try {
+			// Configuration
 			prop.load(new FileInputStream("config.properties"));
 
+			// ouverture du socket
 			this.connexion = new ServerSocket(Integer.parseInt(prop.getProperty("port")));
 		}
 		catch (IOException e)
@@ -26,8 +48,12 @@ public class ServeurHaut
 		}
 	}
 
+	/**
+	 * Appelée pour démarrer le serveur.
+	 */
 	public void call()
 	{
+		// c'est ce thread pool qui s'occupera de lancer les threads clients 
 		ExecutorService res = Executors.newFixedThreadPool(Integer.parseInt(prop.getProperty("coMax")));
 		
 		try 
@@ -36,11 +62,13 @@ public class ServeurHaut
 			{
 				Socket client = connexion.accept();
 
-				client.setSoTimeout(Integer.parseInt(prop.getProperty("port")) * 1000);
+				client.setSoTimeout(Integer.parseInt(prop.getProperty("temps")) * 1000);
 				
 				System.out.println("Client connecté : " + client);
 				
+				// On crée le thread client...
 				GestionnaireClientHaut handler = new GestionnaireClientHaut(client);
+				// et au lieu de le 'start' manuellement, on le délègue au ThreadPool 
 				res.submit(handler);
 			}		
 		}
